[ToggleButtons] Move from lab to core
