+++ docs/src/pages.js
@@ -39,12 +39,13 @@ const pages = [
           { pathname: '/components/floating-action-button' },
           { pathname: '/components/pickers', title: 'Date / Time' },
           { pathname: '/components/radio-buttons' },
+          { pathname: '/components/rating' },
           { pathname: '/components/selects' },
           { pathname: '/components/slider' },
-          { pathname: '/components/rating' },
           { pathname: '/components/switches' },
           { pathname: '/components/text-fields' },
           { pathname: '/components/transfer-list' },
+          { pathname: '/components/toggle-button' },
         ],
       },
       {
@@ -126,7 +127,6 @@ const pages = [
           { pathname: '/components/rating' },
           { pathname: '/components/slider-styled' },
           { pathname: '/components/timeline' },
-          { pathname: '/components/toggle-button' },
           { pathname: '/components/trap-focus' },
           { pathname: '/components/tree-view' },
         ],
