+++ docs/src/pages/components/toggle-button/StandaloneToggleButton.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import CheckIcon from '@material-ui/icons/Check';
-import ToggleButton from '@material-ui/lab/ToggleButton';
+import ToggleButton from '@material-ui/core/ToggleButton';
 
 export default function StandaloneToggleButton() {
   const [selected, setSelected] = React.useState(false);
