+++ packages/material-ui/src/ToggleButton/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './ToggleButton';
+export * from './ToggleButton';
