+++ packages/material-ui/src/ToggleButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './ToggleButton';
