+++ packages/material-ui/src/ToggleButtonGroup/ToggleButtonGroup.js
@@ -1,0 +1,133 @@
+import * as React from 'react';
+import { isFragment } from 'react-is';
+import PropTypes from 'prop-types';
+import clsx from 'clsx';
+import { withStyles } from '@material-ui/core/styles';
+import { capitalize } from '@material-ui/core/utils';
+import isValueSelected from './isValueSelected';
+
+export const styles = (theme) => ({
+  /* Styles applied to the root element. */
+  root: {
+    display: 'inline-flex',
+    borderRadius: theme.shape.borderRadius,
+  },
+  /* Styles applied to the root element if `orientation="vertical"`. */
+  vertical: {
+    flexDirection: 'column',
+  },
+  /* Styles applied to the children. */
+  grouped: {},
+  /* Styles applied to the children if `orientation="horizontal"`. */
+  groupedHorizontal: {
+    '&:not(:first-child)': {
+      marginLeft: -1,
+      borderLeft: '1px solid transparent',
+      borderTopLeftRadius: 0,
+      borderBottomLeftRadius: 0,
+    },
+    '&:not(:last-child)': {
+      borderTopRightRadius: 0,
+      borderBottomRightRadius: 0,
+    },
+  },
+  /* Styles applied to the children if `orientation="vertical"`. */
+  groupedVertical: {
+    '&:not(:first-child)': {
+      marginTop: -1,
+      borderTop: '1px solid transparent',
+      borderTopLeftRadius: 0,
+      borderTopRightRadius: 0,
+    },
+    '&:not(:last-child)': {
+      borderBottomLeftRadius: 0,
+      borderBottomRightRadius: 0,
+    },
+  },
+});
+
+const ToggleButtonGroup = React.forwardRef(function ToggleButtonGroup(props, ref) {
+  const {
+    children,
+    classes,
+    className,
+    exclusive = false,
+    onChange,
+    orientation = 'horizontal',
+    size = 'medium',
+    value,
+    ...other
+  } = props;
+
+  const handleChange = (event, buttonValue) => {
+    if (!onChange) {
+      return;
+    }
+
+    const index = value && value.indexOf(buttonValue);
+    let newValue;
+
+    if (value && index >= 0) {
+      newValue = value.slice();
+      newValue.splice(index, 1);
+    } else {
+      newValue = value ? value.concat(buttonValue) : [buttonValue];
+    }
+
+    onChange(event, newValue);
+  };
+
+  const handleExclusiveChange = (event, buttonValue) => {
+    if (!onChange) {
+      return;
+    }
+
+    onChange(event, value === buttonValue ? null : buttonValue);
+  };
+
+  return (
+    <div
+      role="group"
+      className={clsx(
+        classes.root,
+        {
+          [classes.vertical]: orientation === 'vertical',
+        },
+        className,
+      )}
+      ref={ref}
+      {...other}
+    >
+      {React.Children.map(children, (child) => {
+        if (!React.isValidElement(child)) {
+          return null;
+        }
+
+        if (process.env.NODE_ENV !== 'production') {
+          if (isFragment(child)) {
+            console.error(
+              [
+                "Material-UI: The ToggleButtonGroup component doesn't accept a Fragment as a child.",
+                'Consider providing an array instead.',
+              ].join('\n'),
+            );
+          }
+        }
+
+        return React.cloneElement(child, {
+          className: clsx(
+            classes.grouped,
+            classes[`grouped${capitalize(orientation)}`],
+            child.props.className,
+          ),
+          onChange: exclusive ? handleExclusiveChange : handleChange,
+          selected:
+            child.props.selected === undefined
+              ? isValueSelected(child.props.value, value)
+              : child.props.selected,
+          size: child.props.size || size,
+        });
+      })}
+    </div>
+  );
+});
