+++ packages/material-ui/src/ToggleButtonGroup/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './ToggleButtonGroup';
+export * from './ToggleButtonGroup';
