+++ packages/material-ui/src/ToggleButtonGroup/index.js
@@ -1,0 +1,1 @@
+export { default } from './ToggleButtonGroup';
