+++ packages/material-ui/src/ToggleButtonGroup/isValueSelected.js
